Restructure package and fix imports
