+++ AI_Desktop_App/models/image_generator.py
