+++ AI_Desktop_App/models/text_generator.py
