+++ AI_Desktop_App/utils/config.py
